backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.24.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
rasterio>=1.3.0
Pillow>=9.0.0
huggingface_hub